deploying_service/deploying_pytorch/convert_openvino/requirements.txt
torch==1.11.0
torchvision==0.12.0
onnx==1.9.0
onnxruntime==1.8.0
protobuf==3.19.4
openvino-dev==2022.1.0